SyntenyQC_pkg_1.6post2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "SyntenyQC"
version = "1.6post2"
requires-python = "<= 3.12.9,>=3.10.0"
dependencies = [
  "plotly>=6.0.0,<7.0.0",
  "biopython>=1.80, <=1.83",
  "networkx>=3.0.0,<4.0.0",
  "pandas>=2.0,<3.0"
]
authors = [{name = "Timothy Kirkwood", email = "tdjkirkwood@hotmail.com"}]
maintainers = [{name = "Timothy Kirkwood", email = "tdjkirkwood@hotmail.com"}]
description = "A command line app for the pre-processing and quality control of synteny plot neighbourhoods"  
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["bgc", "biosynthetic gene cluster", "synteny", "synteny plot", "genome neighbourhood", "neighbourhood"]
classifiers = ["Development Status :: 4 - Beta",
  	       "Programming Language :: Python :: 3",
	       "License :: OSI Approved :: MIT License",
               "Operating System :: Microsoft :: Windows",
               "Intended Audience :: Science/Research",
               "Topic :: Scientific/Engineering :: Bio-Informatics"]

[project.urls]
Homepage = "https://github.com/Tim-Kirkwood/SyntenyQC"
Issues = 'https://github.com/Tim-Kirkwood/SyntenyQC/issues'

[project.scripts]
SyntenyQC = "SyntenyQC.app:main_cli"